Separar lógica relacionada con la homoclave y agrego los mismos tests usados en RFCFacil.js
